Make them look a little better :D

+++ src/entities/World.java
@@ -2,7 +2,7 @@ package entities;
 
 public class World {
 	
-	public double width = 1024;
-	public double height = 800;
+	public int width = 1024;
+	public int height = 800;
 	
 }
